Detect failed tests in ARM CI
This commit fixes ARM testing so that it reports a failure when unit
tests fail. This was broken because `find -exec` was not doing what
was expected. To fix this, we now use `find | xargs`.

See https://apple.stackexchange.com/questions/49042/how-do-i-make-find-fail-if-exec-fails.
`find -exec` returns a 0 exit code unless an error occurred while
traversing the directories; the return status of executed commands does
not matter.

Go versions < 1.9 are no longer tested on ARM due to an issue with QEMU
user-level emulation. This issue was fixed for Go 1.9 in
https://github.com/golang/go/commit/2673f9e.